Rework branch list into multi-feature form (#48)
https://github.com/user-attachments/assets/8a6f1a0f-1bbd-4667-b7ee-60ef5211bd91

This PR fleshes out the branch switcher by turning the list into a radio form with multiple buttons for actions.

The checkout behavior has been moved into a "Checkout" button at the bottom of the form alongside new "Delete" and "Force Delete" buttons. Clicking on the name of a branch now selects the radio button associated with it.